move _add_processing_info to util.decorator
