+++ obspy/core/trace.py
@@ -13,21 +13,18 @@ from __future__ import (absolute_import, division, print_function,
 from future.builtins import *  # NOQA
 from future.utils import native_str
 
-import importlib
-import inspect
 import math
 import warnings
 from copy import copy, deepcopy
 
 import numpy as np
-from decorator import decorator
 
 from obspy.core import compatibility
 from obspy.core.utcdatetime import UTCDateTime
 from obspy.core.util import AttribDict, create_empty_data_chunk
 from obspy.core.util.base import _get_function_from_entry_point
 from obspy.core.util.decorator import (
-    deprecated, raise_if_masked, skip_if_no_data)
+    _add_processing_info, deprecated, raise_if_masked, skip_if_no_data)
 from obspy.core.util.misc import (flat_not_masked_contiguous, get_window_times,
                                   limit_numpy_fft_cache)
 
@@ -205,46 +202,6 @@ class Stats(AttribDict):
 
     def _repr_pretty_(self, p, cycle):
         p.text(str(self))
-
-
-@decorator
-def _add_processing_info(func, *args, **kwargs):
-    """
-    This is a decorator that attaches information about a processing call as a
-    string to the Trace.stats.processing list.
-    """
-    callargs = inspect.getcallargs(func, *args, **kwargs)
-    callargs.pop("self")
-    kwargs_ = callargs.pop("kwargs", {})
-    module = inspect.getmodule(func).__name__
-    if '.' in module:
-        module = module.split('.')[0]
-    version = getattr(importlib.import_module(module), '__version__')
-    if module == 'obspy':
-        module = 'ObsPy'
-    info = "{module} {version}: {function}(%s)".format(
-        module=module,
-        version=version,
-        function=func.__name__)
-    arguments = []
-    arguments += \
-        ["%s=%s" % (k, repr(v)) if not isinstance(v, native_str) else
-         "%s='%s'" % (k, v) for k, v in callargs.items()]
-    arguments += \
-        ["%s=%s" % (k, repr(v)) if not isinstance(v, native_str) else
-         "%s='%s'" % (k, v) for k, v in kwargs_.items()]
-    arguments.sort()
-    info = info % "::".join(arguments)
-    self = args[0]
-    result = func(*args, **kwargs)
-    # Attach after executing the function to avoid having it attached
-    # while the operation failed.
-    try:
-        self._internal_add_processing_info(info)
-    except AttributeError:
-        for tr in self:
-            tr._internal_add_processing_info(info)
-    return result
 
 
 class Trace(object):
